==> For Allison/01.Get_CaliforniaSites.ipynb <==
## Code to convert text file with gauges to .csv format. + Instruction to download gauge numbers' data.

This notebook contains a code to convert .txt file with site numbers and names of gauges obtained on the usgs.gov website to .csv format. The output file "california gages.csv" is often used in the other notebooks of this project to access the list of gauge numbers.

Steps to obtain the input file in .txt format:
* Go to https://waterdata.usgs.gov/nwis
* At the top right corner, for "Data Category" select "Surface Water", and for "Geographic Area" select the state needed.
* After getting to "USGS Surface-Water Data for [state of choice]" page, choose the "Field Measurements" button from the options available. 
* Choose site selection criteria. Check "Number of observations" under Data Attribute only. Submit.
* Filter sites: under "Number of observations" type 500 (or another desired minimum of field measurements observations).
* Choose output format: select "table format" in "Site-description information displayed in ".
*  Select fields to include in the file. Here, we only care about the Site identification number, any other columns are optional.
* In "Site-description information displayed in " select "  --saved to file". Submit.
* When prompted to choose the directory to download the file, choose the directory of this project and make sure it is in the same folder as this notebook.
* Open the file in any text editor and delete the comments on top of the file before the column names "agency_cd	site_no	station_nm".

The code below takes that file and converts it to a workable csv format.

import pandas as pd
import requests
from bs4 import BeautifulSoup
import time

file = '500PtSitesCAGauges.txt'
df = pd.read_csv(file, delimiter = "/")
df.to_csv("california gages.csv", index=False)

This code accesses the site's info pages and prints out the statement about whether it exists. Not essential to the project but might be helpful to confirm that the link works before proceeding to the next steps.

url = "https://waterdata.usgs.gov/nwis/inventory?agency_code=USGS&site_no="
pot = []
i = 0
for site in df["site_no"]:
  url = url + str(site)
  req = requests.get(url)
  url = url[:-len(str(site))]
  soup = BeautifulSoup(req.content, 'html.parser')
  pot.insert(i,soup)
  i = i+1
  time.sleep(0.2)
  if req.status_code == 200:
    print('Web site exists')
  else:
    print('Web site does not exist')
