==> For Allison/02.Download_field_measurements.ipynb <==
## Code to collect field measurements data

This notebook contains a code to scrape field measurements data from usgs.gov and store each gauge's data in its own .csv file inside field_measurement_data folder. To obtain a final database of all measurements go to field_measurement_data folder --> 03.combine_field_data.ipynb

import time
import requests
import pandas as pd
import numpy as np
import bs4
import csv
import re

cal_gages = pd.read_csv('california gages.csv', encoding='latin-1', low_memory=False)

cal_gages

for col in cal_gages.columns:
    print(col)

cal_gages["site_no"]

### CSV Maker
This code will go in a loop to read each entry in caliornia_gages.csv and extract field measurements.

url -> to be replaced by an entry's *modified* url

site_no -> to be replaced by an entry's site number

for site_no in cal_gages["site_no"].values[1:]:
    url = "https://waterdata.usgs.gov/nwis/measurements?site_no=" + site_no + "&agency_cd=USGS&format=rdb_expanded"
    response = requests.get(url)
    if response.status_code == 200:
        print(site_no + " Success")
        data = pd.read_csv(url, sep='\t', lineterminator='\n', comment='#', header=0, skiprows=[29])
        data.to_csv("field_measurement_data/" + site_no + ".csv", index=False)
    else:
        print(site_no + " Error: ",response.status_code)
