==> For Allison/04.Download_water-yr_summary.ipynb <==
## Code to obtain Water-year summary data for all gauges and put it in one dataset

This notebook contains a script to scrape water-year summary data from the usgs.gov website. It works by accessing web pages of each gauge's water-year summary and extracting the text we need (Remarks, Gage, Location) from their HTML tags. The code is easy to modify in case some other states have different formatting of the water-year summary page. This script worked great with California, Washington, Oregon, Arizona, and Nevada states.

After the scripting, there is a code to preprocess data by removing extra tags that were pulled with content during scraping. The output of that is clean_csv_ water-year summary.csv which will be used in water_year_summary_model.ipynb to label each gauge as valid/invalid.

from bs4 import BeautifulSoup
import requests
import pandas as pd
from unicodedata import normalize
import time
from random import randrange

# Import the list of gauges; remove the first row that contains column titles (specific to this dataset).
cal_gages = pd.read_csv('california gages.csv', encoding='latin-1', low_memory=False)
cal_gages = cal_gages.iloc[1:, :]

cal_gages

# Scraping code.
# Goes through water-year summary page for each gauge, scraps the relevant text about Location, Gage and Remarks;
# saves the array into a dataframe and outputs it to a .csv file.
# There is a pause between scrapping each page (from 10 to 20 seconds, randomly) to avoid getting blocked by the website.

data = {"Site No": [], "Location":[], "Gage": [], "Remarks": []}

for entry in cal_gages["site_no"] :
    
    sec = randrange(10, 20)
    site_no = entry
    url = "https://waterdata.usgs.gov/nwis/wys_rpt/?site_no=" + site_no
    time.sleep(sec)
    req = requests.get(url)
    summary = req.text
    soup = BeautifulSoup(summary)
    
    try:
        location = soup.find("div", {"id":"location"}).get_text()
        location = normalize('NFKD', location)
    except AttributeError: #this part is necessary when a searched text does not exist on the page
        location = ""      #create NA in data to avoid errors
    
    try:
        gage = soup.find("div", {"id":"gage"}).get_text()
        gage = normalize('NFKD', gage)
    except AttributeError:
        gage = ""
    
    try:
        remarks = soup.find("div", {"id":"remarks"}).get_text()
        remarks = normalize('NFKD', remarks)
    except AttributeError:
        remarks = ""
    
    data["Site No"].append(site_no)
    data["Location"].append(location)
    data["Gage"].append(gage)
    data["Remarks"].append(remarks)

df_scraped = pd.DataFrame(data)
df_scraped.to_csv('raw_ water-year summary.csv', index=False)

#Save raw data in .csv
df = pd.read_csv('raw_ water-year summary.csv')

df

df.info()

df_clean = df

# Remove redundant titles in columns: "LOCATION - ", "GAGE - ", "REMARKS - ".

df_clean['Location'] = df_clean['Location'].astype(str).str.replace('LOCATION - ', '')
df_clean['Gage'] = df_clean['Gage'].astype(str).str.replace('GAGE - ', '')
df_clean['Remarks'] = df_clean['Remarks'].astype(str).str.replace('REMARKS - ', '')

df_clean

# Save cleaned dataframe in .csv
df_clean.to_csv('clean_csv_ water-year summary.csv', index=False)
