==> For Allison/06.Download_rating_curve_data.ipynb <==
# Code to collect Rating Curve data
- Scraping of the rating curve of stream gauge data provided by USGS. 
#### Input:
Set of stream gages.
#### Output:
Rating curve data for every site. Formated where all data is stored in one data frame that can be accessed by inputting the specific gage.
#### Implementation:
Scraping is done by using beautifulsoup, requests staggered such to not overwhelem the system, data 

import pandas as pd
import numpy as np
import requests
import time
from bs4 import BeautifulSoup

# Just the site number is needed
ca = pd.DataFrame(pd.read_csv('CA_valid_gages.csv')["Site No"])

ratings = pd.DataFrame()
noPageGage = pd.DataFrame()
noPageGage['Site No'] = []
for gageNum in ca['Site No']:
    try:
        url = ('https://nwis.waterdata.usgs.gov/nwisweb/data/ratings/exsa/USGS.' + str(gageNum) + '.exsa.rdb')
        response = requests.get(url)
        try:
            if response.status_code == 404:
                noPageGage.loc[len(noPageGage)] = {'Site No': str(gageNum)}
                continue;
        except:
            pass
        # Processing data from request
        soup = BeautifulSoup(response.text, 'html.parser')
        lines = soup.string.split('\n')
        lines = [line.strip() for line in lines if not line.strip().startswith('#')]
        data = [line.split("\t") for line in lines]
        df = pd.DataFrame(data[1:], columns=data[0])
        df['Site No'] = str(gageNum)
        col = df.pop('Site No')
        df.insert(0, 'Site No', col)
        ratings = pd.concat([ratings,df], ignore_index = True)
        time.sleep(3)
    except requests.exceptions.HTTPError as e:
        # Handle other HTTP errors here
        print(f"HTTP error {e.response.status_code} occurred: {url}")

ratings.to_csv('rating_curve_wa.csv', index=False)

ratings

print(str(ratings[ratings["INDEP"].isna()].size) + '/' + str(ratings.size))
print(str(ratings[ratings["SHIFT"].isna()].size) + '/' + str(ratings.size))
print(str(ratings[ratings["DEP"].isna()].size) + '/' + str(ratings.size))
print(str(ratings[ratings["STOR"].isna()].size) + '/' + str(ratings.size))

ratings.describe()
